Move Enroll table from core to student_core
