+++ core/admin.py
@@ -28,7 +28,3 @@ class ResourceSectionAdmin(admin.ModelAdmin):
 @admin.register(Resource)
 class ResourceAdmin(admin.ModelAdmin):
     search_fields = ['id', 'name', 'section__name']
-
-@admin.register(Enroll)
-class Enroll(admin.ModelAdmin):
-    search_fields = ['studentUserID', 'classroom', 'studentIndex' ]
